Add asyncOptions setting to enable passport asynchronous options feature. Updated config get call to support 'authentication' and 'auth' settings. Eliminated need for babel transpiling and adjusted code pattern to more closely match authentication-local.
