--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "plugins": [ "add-module-exports" ],
-  "presets": [ "es2015" ]
-}
