+++ .gitignore
@@ -29,9 +29,6 @@ node_modules
 # Users Environment Variables
 .lock-wscript
 
-# The compiled/babelified modules
-lib/
-
 # IDE
 .idea
