+++ lib/index.js
@@ -1,8 +1,9 @@
-import Debug from 'debug';
-import merge from 'lodash.merge';
-import omit from 'lodash.omit';
-import DefaultVerifier from './verifier';
-import { Strategy as LdapStrategy } from 'passport-ldapauth';
+const Debug = require('debug');
+const merge = require('lodash.merge');
+const omit = require('lodash.omit');
+const DefaultVerifier = require('./verifier');
+
+const passportLdap = require('passport-ldapauth');
 
 const debug = Debug('feathers-authentication-ldap');
 const defaults = {
@@ -26,26 +27,32 @@ class JWTVerifier {
   }
 }
 
-// Export ldap-auth init function
-export default function init (options = {}) {
+function init (options = {}) {
   return function ldapAuth () {
     const app = this;
     const _super = app.setup;
 
     if (!app.passport) {
-      throw new Error(`Can not find app.passport. Did you initialize feathers-authentication before feathers-authentication-ldap?`);
+      throw new Error(`Can not find app.passport. Did you initialize feathers-authentication before @feathersjs/authentication-ldap?`);
     }
 
-    // Construct ldapSettings for passport ldap strategy
     let name = options.name || defaults.name;
-    let authOptions = app.get('auth') || {};
+    let authOptions = app.get('authentication') || app.get('auth') || {};
     let ldapOptions = authOptions[name] || {};
-    const ldapSettings = merge({}, defaults, ldapOptions, omit(options, ['Verifier']));
-    const Verifier = options.Verifier || DefaultVerifier;
+    const ldapSettings = merge({}, defaults, ldapOptions, omit(options, ['Verifier, AsyncOptions']));
+    const asyncOptions = options.asyncOptions || null;
+    let Verifier = DefaultVerifier;
 
-    // plugin setup: register strategy in feathers passport
+    if (options.Verifier) {
+      Verifier = options.Verifier;
+    }
+
+    // function to get dynamic settings when using asyncOptions in strategy
+    const getLdapSettings = (req, passportCallback) => {
+      passportCallback(null, merge({}, ldapSettings, asyncOptions(req)));
+    };
+
     app.setup = function () {
-      // be sure feathers setup was called
       let result = _super.apply(this, arguments);
       let verifier = new Verifier(app, ldapSettings);
 
@@ -53,9 +60,8 @@ export default function init (options = {}) {
         throw new Error(`Your verifier must implement a 'verify' function. It should have the same signature as function(request, user, done)`);
       }
 
-      // Register 'ldap' strategy with passport
       debug('Registering ldap authentication strategy with options:', ldapSettings);
-      app.passport.use(ldapSettings.name, new LdapStrategy(ldapSettings, verifier.verify.bind(verifier)));
+      app.passport.use(ldapSettings.name, new passportLdap.Strategy((asyncOptions ? getLdapSettings : ldapSettings), verifier.verify.bind(verifier)));
       app.passport.options(ldapSettings.name, ldapSettings); // do we need this ??
 
       return result;
@@ -63,8 +69,11 @@ export default function init (options = {}) {
   };
 }
 
+module.exports = init;
+
 // Exposed Modules
-Object.assign(init, {
+Object.assign(module.exports, {
+  default: init,
   defaults,
   Verifier: DefaultVerifier,
   JWTVerifier
