+++ package.json
@@ -28,19 +28,16 @@
     "node": ">= 4.6.0"
   },
   "scripts": {
-    "prepublish": "npm run compile",
     "publish": "git push origin --tags && npm run changelog && git push origin",
     "release:patch": "npm version patch && npm publish",
     "release:minor": "npm version minor && npm publish",
     "release:major": "npm version major && npm publish",
     "changelog": "github_changelog_generator && git add CHANGELOG.md && git commit -am \"Updating changelog\"",
-    "compile": "rimraf lib/ && babel -d lib/ src/",
-    "watch": "babel --watch -d lib/ src/",
-    "lint": "semistandard src/**/*.js test/**/*.js --fix",
+    "lint": "semistandard --fix",
     "mocha": "mocha --opts mocha.opts",
-    "coverage": "istanbul cover _mocha -- --opts mocha.opts",
-    "test": "npm run compile && npm run lint && npm run coverage",
-    "start": "npm run compile && node example/app"
+    "coverage": "istanbul cover node_modules/mocha/bin/_mocha -- --opts mocha.opts",
+    "test": "npm run lint && npm run coverage",
+    "start": "node example/app"
   },
   "semistandard": {
     "sourceType": "module",
@@ -58,10 +55,6 @@
     "passport-ldapauth": "^2.0.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-core": "^6.20.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.18.0",
     "body-parser": "^1.15.2",
     "chai": "^4.1.0",
     "feathers-authentication": "^1.0.0",
@@ -73,7 +66,6 @@
     "istanbul": "^1.1.0-alpha.1",
     "sinon": "^4.0.0",
     "mocha": "^5.0.0",
-    "rimraf": "^2.5.4",
     "semistandard": "^12.0.0",
     "sinon-chai": "^2.8.0"
   }
